Fix accounts creation when accounts are available
