--- app/components/modals/AddMixerAccountsModal/Modal.js
@@ -1,39 +1,0 @@
-import { FormattedMessage as T } from "react-intl";
-import { TextInput, PassphraseModalField } from "inputs";
-import { PassphraseModal } from "../PassphraseModal";
-
-const Modal = ({
-  mixedAccountName, changeAccountName, setMixedAccountName, setChangeAccountName,
-  ...props
-}) => (
-  <PassphraseModal {...{ ...props }}>
-    <PassphraseModalField
-      label={<T id="addMixerAccountModal.mixedAccountName" m="Mixed Account Name" />}
-    >
-      <TextInput
-        autoFocus
-        required
-        id="name"
-        type="text"
-        placeholder="Mixed Account Name"
-        value={mixedAccountName}
-        onChange={(e) => setMixedAccountName(e.target.value)}
-      />
-    </PassphraseModalField>
-    <PassphraseModalField
-      label={<T id="addMixerAccountModal.changeAccountName" m="Change Account Name" />}
-    >
-      <TextInput
-        autoFocus
-        required
-        id="name"
-        type="text"
-        placeholder="Change Account Name"
-        value={changeAccountName}
-        onChange={(e) => setChangeAccountName(e.target.value)}
-      />
-    </PassphraseModalField>
-  </PassphraseModal>
-);
-
-export default Modal;
