+++ app/components/views/AccountsPage/Privacy/ConfigMixer.js
@@ -4,11 +4,14 @@ import { AddMixerAccountsModal } from "modals";
 import { WatchOnlyWarnNotification, Subtitle } from "shared";
 import { MIXED_ACCOUNT, CHANGE_ACCOUNT } from "constants";
 import { useState, useEffect } from "react";
+import { useDispatch } from "react-redux";
+import * as amc from "actions/AccountMixerActions";
 
 function ConfigMixer({
   isCreateAccountDisabled, accounts
 }) {
-  const [areAccountsAvailable, setAreAvailable] = useState(null);
+  const dispatch = useDispatch();
+  const [areAccountsAvailable, setAreAvailable] = useState(false);
 
   const checkAvailableAccounts = () => {
     const mixedExists = accounts.find(({ accountName }) => accountName === MIXED_ACCOUNT);
@@ -16,7 +19,20 @@ function ConfigMixer({
 
     return !mixedExists && !changeExists;
   }
-  useEffect(() => setAreAvailable(checkAvailableAccounts()), []);
+  const [ mixedAccountName, setMixedAccountName ] = useState("");
+  const [ changeAccountName, setChangeAccountName ] = useState("");
+  const isValid = () => !(!mixedAccountName || !changeAccountName);
+  const onSubmit = (passphrase) => {
+    dispatch(amc.createNeededAccounts(passphrase, mixedAccountName, changeAccountName));
+  }
+
+  useEffect(() => {
+    if (checkAvailableAccounts()) {
+      setAreAvailable(true)
+      setMixedAccountName(MIXED_ACCOUNT);
+      setChangeAccountName(CHANGE_ACCOUNT);
+    }
+  }, []);
 
   return (
     <>
@@ -25,7 +41,9 @@ function ConfigMixer({
         {
           areAccountsAvailable ? (
             <div>
-              <T id="privacy.create.default.title" m="Create Default Accounts" />
+              <div>
+                <T id="privacy.create.default.title" m="Create Default Accounts" />
+              </div>
               <T id="privacy.create.default.description"
                 m={`If continue the accounts {mixed} and {change} are going to be created.
                   Which are the default ones for the mixer. {boldMessage}`}
@@ -38,6 +56,18 @@ function ConfigMixer({
                     </span>
                 }}
               />
+              <div>
+                <WatchOnlyWarnNotification isActive={isCreateAccountDisabled}>
+                  <PassphraseModalButton
+                    {...{
+                      onSubmit, mixedAccountName, changeAccountName, setMixedAccountName,
+                      setChangeAccountName, disabled: isCreateAccountDisabled
+                    }}
+                    modalTitle={<T id="accounts.defaultAccountConfirmations" m="Create default accounts" />}
+                    buttonLabel={<T id="accounts.addNewButton" m="Create Needed Accounts" />}
+                  />
+                </WatchOnlyWarnNotification>
+              </div>
             </div>
           ) : (
               <>
@@ -61,6 +91,10 @@ function ConfigMixer({
                 </div>
                 <WatchOnlyWarnNotification isActive={isCreateAccountDisabled}>
                   <PassphraseModalButton
+                    {...{
+                      onSubmit, mixedAccountName, changeAccountName, setMixedAccountName,
+                      setChangeAccountName, disabled: isCreateAccountDisabled
+                    }}
                     disabled={isCreateAccountDisabled}
                     modalTitle={<T id="accounts.newAccountConfirmations" m="Create needed accounts" />}
                     modalComponent={AddMixerAccountsModal}
